+++ src/ui/components/modals/index.ts
@@ -5,4 +5,5 @@ export { default as InfoModal } from './InfoModal';
 export { default as LogoutModal } from './LogoutModal';
 export { default as VisibilityModal } from './VisibilityModal';
 export { default as SortModal } from './SortModal';
+export { ChangeVisibilityModal } from './ChangeVisibilityModal';